+++ src/main/java/io/github/meechapooch/chatplusplus/chat/ChatThing.java
@@ -1,0 +1,7 @@
+package io.github.meechapooch.chatplusplus.chat;
+
+public interface ChatThing {
+    void sendMessage(String sender, String message);
+
+    String getName();
+}
