--- src/main/java/to/us/ravinesquaad/plugins/chatplusplus/ChatThing.java
@@ -1,6 +1,0 @@
-package to.us.ravinesquaad.plugins.chatplusplus;
-
-public interface ChatThing {
-    public void sendMessage(String sender, String message);
-    public String getName();
-}
