--- src/main/java/to/us/ravinesquaad/plugins/chatplusplus/RespondCompleter.java
@@ -1,14 +1,0 @@
-package to.us.ravinesquaad.plugins.chatplusplus;
-
-import org.bukkit.command.Command;
-import org.bukkit.command.CommandSender;
-import org.bukkit.command.TabCompleter;
-
-import java.util.List;
-
-public class RespondCompleter implements TabCompleter {
-    @Override
-    public List<String> onTabComplete(CommandSender sender, Command command, String alias, String[] args) {
-        return null;
-    }
-}
