--- src/main/java/to/us/ravinesquaad/plugins/chatplusplus/Test.java
@@ -1,11 +1,0 @@
-package to.us.ravinesquaad.plugins.chatplusplus;
-
-public class Test {
-    public static void main(String... args) {
-        int i;
-        for (i = 0; i < 10; i++) {
-            if(i==5) break;
-        }
-        System.out.println(i);
-    }
-}
